+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Added 26.2 support
   - This is the last release that will also support 1.21.11
   - Added an alert for users on 1.21.11  
+- Added Frogcoin blessing display
 
 ### Changes
 - Made the bestiary display show in alphabetical order